fix(esc): use of injected component pattern for different ctas
